opendingux: big installer overhaul
Big rework on the installer. It can now be used to flash many more
devices, including the GCW-Zero, the RG-350(M), and the supported JZ4760
based devices like the RS-97 v2.0.

To build:
CONFIG=installer ./rebuild.sh

the output/installer/images/vmlinuz.bin can then be used with
odboot-client (https://github.com/opendingux/odbootd) along with the
latest update OPK (which can be a nightly build from
http://od.abstraction.se/opendingux/) in order to flash a device from
USB.

+++ rebuild.sh
@@ -3,7 +3,7 @@ set -e
 
 if [ -z "$CONFIG" ] ; then
 	echo "\$CONFIG not set. Please set it to the variant to build."
-	echo "Valid values are: gcw0, rs90, lepus"
+	echo "Valid values are: gcw0, rs90, lepus, installer"
 	exit 1
 fi
 
@@ -14,16 +14,20 @@ rm -rf output/${CONFIG}
 # Use the default config.
 make od_${CONFIG}_defconfig BR2_EXTERNAL=board/opendingux O=output/${CONFIG}
 
+[ "${CONFIG}" = "installer" ] || SDK_TARGET=sdk
+
 # Perform the build.
 echo "Starting build..."
-nice make sdk BR2_SDK_PREFIX=${CONFIG}-toolchain O=output/${CONFIG}
+nice make ${SDK_TARGET} BR2_SDK_PREFIX=${CONFIG}-toolchain O=output/${CONFIG}
 
-echo "Recompressing SDK to XZ..."
-ARCHIVE_NAME=opendingux-${CONFIG}-toolchain.`date +'%Y-%m-%d'`
-gzip -d -c output/${CONFIG}/images/${CONFIG}-toolchain.tar.gz | xz -T0 -9 > output/${CONFIG}/images/$ARCHIVE_NAME.tar.xz
-rm output/${CONFIG}/images/${CONFIG}-toolchain.tar.gz
+if [ "${SDK_TARGET}" ] ; then
+	echo "Recompressing SDK to XZ..."
+	ARCHIVE_NAME=opendingux-${CONFIG}-toolchain.`date +'%Y-%m-%d'`
+	gzip -d -c output/${CONFIG}/images/${CONFIG}-toolchain.tar.gz | xz -T0 -9 > output/${CONFIG}/images/$ARCHIVE_NAME.tar.xz
+	rm output/${CONFIG}/images/${CONFIG}-toolchain.tar.gz
 
-echo "The SDK has been built at:"
-echo "output/${CONFIG}/images/$ARCHIVE_NAME.tar.xz"
-echo
-echo "Remember to run ./relocate-sdk.sh after extracting it to your desired location"
+	echo "The SDK has been built at:"
+	echo "output/${CONFIG}/images/$ARCHIVE_NAME.tar.xz"
+	echo
+	echo "Remember to run ./relocate-sdk.sh after extracting it to your desired location"
+fi
